Generate report even if exited with `sys.exit()`
During the scenario run, formatter only caches the data,
that re being compiled to HTML at the end, in `close()`
function. This is unfortunate, if code exits with `sys.exit()`
the report is empty. Fix this with registering `atexit` hook,
which will call `close()` on exit. Make sure, `close()` is
executed at most one, otherwise report is doubled.

Closes #34